+++ src/constants.ts
@@ -1,0 +1,7 @@
+export const ForceLayoutTypeMap: { [key: string]: boolean } = {
+  'gForce': true,
+  'fruchterman': true,
+  'forceAtlas2': true,
+  'force': true,
+  'graphin-force': true,
+}